Begin making Accessor interface more robust
This fixes leaky (and broken) behavior for when nodes are dropped an
should be removed from the hierarchy.  A small base of tests have been
added to cover a few of those cases.

The ability to disable access for swaths of the hierarchy is still
missing, but this should lay the groundwork for more easily adding that
in the future.