Gripe when using non-raw block device
If the S_IFBLK instance of a block device is used, rather than the
S_IFCHR raw instance, flush operations can be very expensive as the OS
is forced to walk the page cache looking for necessary work.

While we will not strictly forbid the operator from choosing to use such
a device, we can at least admonish their choice in the logs.

Fixes #518