phd: Ensure min disk size fits read-only parents
Today PHD's Crucible tests unconditionally specify a 10 GiB guest disk
even if the guest image being used as a read-only parent is larger than
that. To avoid truncating, tests should specify a minimum guest disk
size, and PHD should expand it to the underlying guest image size if
that exceeds what the test specified.

This branch changes the `disk_size_gib` argument to functions that
create Crucible disks to `min_disk_size_gib`. If a disk is created with
a read-only parent, the code that constructs new Crucible disks
(`CrucibleDisk::new`) will now read the FS metadata for the parent image
to determine its size, and uses that to set the size of the created disk
to ensure that the read-only parent image fits.