phd: refactor serial console
Significantly clean up the serial console logic so that it has one task instead
of two and has a trait-based design for processing incoming bytes from the guest
and registering waits for strings. This will make it significantly easier to
implement new kinds of buffers to handle guests that expect different kinds of
serial console implementations.

This commit temporarily removes logging of serial console lines to the primary
test logger (in anticipation of bringing them back later, but logging serial
console output to a file).