add test device for simulating migration failures
Currently, we have no mechanism for testing how a `propolis-server`
process recovers from a failed migration attempt. In order to exercise
this case, this PR introduces a test-only virtual device which can be
configured to fail during migration, either while exporting the device
or while importing it. The device can be configured to fail a set number
of times, to allow tests that force a migration failure and then perform
a subsequent migration that does not fail.

Since a virtual device whose only purpose is to break Propolis seems
like something that we don't want to create in production,
`propolis-server` will only permit the migration-failure test device to
be configured in the TOML config file, and it is *not* added to the
`InstanceSpec` public API. `propolis-server` also ignores any
migration-failure devices configured in the TOML if it was built with
the "omicron-build" feature, so we won't accidentally inject failures
into release builds.