PHD: add `cargo phd` to make using PHD nicer
Human use of `phd-runner` has a few pain points. In particular:

- Because `phd-runner` takes the path to the `propolis-server` binary as
  a CLI argument, it's easy to forget to rebuild `propolis-server` after
  making a change, and accidentally run the tests against a stale
  artifact. This has bitten me a few times, and in some cases, has
  resulted in a non-trivial time waste, both because running the tests
  takes a bit and because I've spent time debugging why the tests were
  still failing after I fixed something, only to discover that I was
  testing the old binary.
- A whole bunch of command-line arguments have to be passed in every
  time, which is a bit annoying to remember.
- The `phd-quickstart.sh` script can be used to avoid typing out some of
  the CLI args, but its management of the temp dir and artifact store is
  not particularly sophisticated. In particular, it uses the same
  directory for both,[^1] making it difficult to reuse the artifact
  store to avoid downloading the same artifacts over and over again.

In order to make testing changes locally during development a bit nicer,
I've added a new `cargo phd` Cargo subcommand for running PHD tests,
which should resolve some of these issues. Because `cargo phd`'s
behavior is a bit more opinionated than `phd-runner`, and it has default
behaviors for a bunch of things that `phd-runner` requires be explicitly
specified, I've implemented it as a separate binary that's completely
opt-in. In particular, I don't think it would make sense to change the
`phd-run` CI jobs to use `cargo phd`.

The primary goal of the cargo subcommand is to ensure that the
`propolis-server` binary used for testing is always up-to-date with
local changes. The subcommand shells out to `cargo build`, to ensure
that the binary is up-to-date, and then passes the path to the freshly
built `propolis-server` to `phd-runner`. This relies on cargo's change
detection to avoid rebuilding the binary when the `propolis-server`
codebase hasn't acutally changed.

If the user passes a `--propolis-server-cmd` argument, that value is
used rather than rebuilding the `propolis-server` binary, so `cargo phd`
can still be used to test a `propolis-server` binary that came from
elsewhere.

In addition, `cargo phd` manages temporary dirs and the artifact store a
bit more intelligently. The artifact store and temp dirs for a test run
are `target/phd/artifacts` and `target/phd/tmp/{timestamp}`,
respectively. This lets us create separate temp dirs for each run, to
avoid clobbering logs and stuff from prior runs, but reuse the existing
artifacts we may have downloaded previously, making it faster to run
tests repeatedly. Because we put both directories in `target/`, they
last longer than `/tmp/` would, making the persistent artifact store
more useful, but they're still blown away by `cargo clean` if the
developer wants to free up some space.

I'll admit that I may also have gotten a bit carried away trying to make
the CLI as nice as possible... :)

[^1]: which, unfortunately, is what `phd-runner` will actually *always*
    do, prior to PR #618...