PHD: fix `--artifact-directory` not doing anything (#618)
Currently, the `phd-runner` CLI requires both a `--tmp-directory`
argument and an `--artifact-directory` argument. This should, in theory,
permit the artifact directory to be located outside the temp directory,
if requested by the user.

However, the `--artifact-directory` argument is not actually honored by
PHD, and the artifact store is always constructed in the temp directory.
This is kind of unfortunate, since there are some nice reasons to
separate the artifact store and the temp directory. For example, I might
want to create a unique temp dir every time I run PHD tests, so that I
don't clobber logs from previous test runs. Under the current behavior,
since the artifact store is always located in the temp dir, this would
mean re-downloading artifacts on every run, which seems sad. Also, it's
not great that we require the `--artifact-directory` argument, but then
don't ever actually use its value :)

This commit fixes this issue, by making the `--artifact-directory`
argument actually control the path of the artifact store. Also, I've
made the `--artifact-directory` argument optional, and used the temp dir
as the default if the artifact dir isn't provided. If this behavior is
unwanted, I can always change this back to a required argument.