PHD: add tests for migration of running processes
In real life, it's likely that VM migration may occur while one or more
user processes are running on the guest. We don't have a lot of testing
for what happens to userspace processes during migration. Therefore,
I've added some tests that create a process on the guest, write a bunch
of data into variables in memory, suspend the process, migrate the
guest, and attempt to read the data back out of memory on the new host
and check that it's still correct.

In addition to a simple test that performs one migration, I've also
added tests that use the migration-failure device to force a migration
attempt to fail while a process is running, and then check that the
process hasn't been messed up after a successful second migration
attempt.

The primary intention behind these tests is to ensure that all dirty
pages of the guest's memory are migrated correctly (see #387). With the
current approach to RAM transfer, we don't run into issues with this,
even in the case of a failed migration, because the `RamPushPrePause`
phase always offers *all* guest RAM, regardless of whether it's dirty.
However, if we change the RAM transfer behavior to only offer dirty
pages both in the `RamPushPrePause` and `RamPushPostPause` migration
phases, the migration-failure tests will now fail, because the failed
migration will have cleared dirty bits on any dirty guest pages, and we
now don't transfer the guest's memory correctly on the second migration.
These tests will allow us to verify that a fix for #387, where we no
longer offer all RAM in the `RamPushPrePause` phase, is working
correctly.