PHD: fix run_shell_command with multiline commands (#621)
Currently, the `TestVm::run_shell_command` method in `phd_framework`
fails incorrectly when the shell is multi-line. This is because the
shell will not echo the literal command, but instead prints a character
at the beginning of each continuation line. On Alpine, this is `> `.
Because `run_shell_command` waits for the literal command line to be
echoed, it will never see the input command again if it's more than one
line, because of the addition of these continuation characters by the
shell, and we time out waiting to see the command echoed.

This branch fixes this by adding continuation characters to `echo_cmd`
when the input command is multi-line. I can't promise this works on all
Linux images, since other shells may use a different character for
multi-line continuations, but it works with the Alpine image we're
currently using...