PHD: refactor `raw_buffer` to use `Parser::parse`
Currently, the `RawBuffer` struct in `phd_framework::serial::raw_buffer`
uses the `Parser::parse_as_vec` method on the `termwiz` escape parser.
This is a bit unfortunate, as it allocates a `Vec` of `Action`s every
time we parse escape codes in a line of serial output. We could,
instead, use the `Parser::parse` method, which takes a `FnMut` closure
that's executed for each parsed `Action`, rather than returning a `Vec`
of `Action`s that we then iterate over, except that the methods we call
for those `Action`s mutably borrow the `RawBuffer`...which we can't do
because calling `Parser::parse` mutably borrows the `Parser`, which is a
field on `RawBuffer`. See this comment for details:
https://github.com/oxidecomputer/propolis/pull/627#discussion_r1468257978

In this PR, I've refactored the `RawBuffer` code to make using
`Parser::parse` rather than `parse_as_vec` possible. By splitting out
the fields other than the `Parser` to a separate struct, we can move the
various methods that have to be called based on the `Action` from
methods on `RawBuffer` to methods on the `Inner` struct. Now,
`RawBuffer` is

```rust
pub struct RawBuffer {
    buffer: Inner,
    parser: Parser,
}
```

so we can mutably borrow the `parser` and `buffer` fields separately.
This way, we can call `Parser::parse`, mutably borrowing the `parser`
field but *not* the rest of the buffer. This allows us to happily pass
in a closure that mutably borrows the `buffer` field, because it's not
already borrowed by calling `parse`.