PHD: add tests for migration of running processes (#623)
In real life, it's likely that VM migration may occur while one or more
user processes are running on the guest. We don't have a lot of testing
for what happens to userspace processes during migration. Therefore,
I've added some tests that create a process on the guest, write a bunch
of data into variables in memory, suspend the process, migrate the
guest, and attempt to read the data back out of memory on the new host
and check that it's still correct.

In addition to a simple test that performs one migration, I've also
added tests for correctness in the face of a _failed_ migration attempt.
These tests force a migration attempt to fail while a process is
running, and then check that the process hasn't been messed up after a
successful second migration attempt.

In order facilitate testing migration after a failed migration,  this PR
introduces a test-only virtual device which can be configured to fail
during migration, either while exporting the device or while importing
it. The device can be configured to fail a set number of times, to allow
tests that force a migration failure and then perform a subsequent
migration that does not fail.

Since a virtual device whose only purpose is to break Propolis seems
like something that we don't want to create in production,
`propolis-server` will only permit the migration-failure test device to
be configured in the TOML config file, and it is *not* added to the
`InstanceSpec` public API. `propolis-server` also ignores any
migration-failure devices configured in the TOML if it was built with
the "omicron-build" feature, so we won't accidentally inject failures
into release builds.

The primary intention behind these tests is to ensure that all dirty
pages of the guest's memory are migrated correctly (see #387). With the
current approach to RAM transfer, we don't run into issues with this,
even in the case of a failed migration, because the `RamPushPrePause`
phase always offers *all* guest RAM, regardless of whether it's dirty.
However, if we change the RAM transfer behavior to only offer dirty
pages both in the `RamPushPrePause` and `RamPushPostPause` migration
phases, the migration-failure tests will now fail, because the failed
migration will have cleared dirty bits on any dirty guest pages, and we
now don't transfer the guest's memory correctly on the second migration.
These tests will allow us to verify that a fix for #387, where we no
longer offer all RAM in the `RamPushPrePause` phase, is working
correctly.