PHD: assume specialized Windows images
Change the Windows guest adapters to assume they're operating on a fully-
specialized, post-OOBE guest image. This fixes tests like the `lspci` lifecycle
test that stop and start a VM and assume that it will go through the same boot
sequence on its initial and subsequent attempts to start.

Also fix up the PHD README to describe the WS2019 adapter (missed this in the
change that added the adapter).

Tested with WS2019 and WS2022 guests.