Remove Inventory mechanism
The Inventory abstraction in propolis-lib has outlived its necessity.
Created initially to keep devices in scope after VM initialization, it
was somewhat cumbersome to deal with, and was providing more friction
that value when it came to accessing the contained Entity bits.  With
its removal, the Instance abstraction can go as well, since it was
simply holding the Machine and the Inventory together.  In this new
world, propolis-lib consumers are left to hold Lifecycle (formerly
Entity) trait objects in whatever way seems appropriate to them.

With the Inventory gone, the parent-child relationships possible through
the former Entity trait are also on the chopping block.  The 440FX/PIIX4
chipset devices have been rearranged to more accurately reflect what is
present in "real" devices.