PHD: convert to async (#633)
Convert the PHD framework to run on an async runtime. This makes it much easier
to cancel running tests in response to signals. This, in turn, makes it easier
to drop all of the objects created by a specific test case, which makes it much
safer to add framework primitives that need to be cleaned up in an orderly
fashion (e.g. ZFS-based copy-on-write file-backed guest disks).

Tested via local runs with Alpine and Debian 11 guests.