fall back to copying artifacts that aren't on a ZFS dataset
This keeps tests from falling over if the user chose an artifact directory that
isn't on a ZFS dataset (e.g. everything is in tmpfs).

This should fix PHD CI, but I'm also going to try to make another change to
allow CI runs to take advantage of the cloning feature.