PHD: allow overriding base commit with `xtask phd`
Currently, `cargo xtask phd` accepts a custom `--base-propolis-branch`
argument, and doesn't automatically populate that argument if it's
provided. However, it doesn't support the `--base-propolis-commit` or
`--base-propolis-cmd` arguments at all, and if they are present, `cargo
xtask phd` will still pass `--base-propolis-branch` to `phd-runner`,
causing a conflict between the `--base-propolis-branch` and
`--base-propolis-commit`/`--base-propolis-cmd` args. This means that
it's impossible to use `cargo xtask phd` with an explicit base commit or
local base binary.

This commit fixes that, by changing `cargo xtask phd run` to also look
for those arguments, and skip setting the default
`--base-propolis-branch` if either are present.