PHD: don't require `--` in xtask arg passthrough
Currently, `cargo xtask phd` requires a `--` delimiter prior to any
arguments that are passed through to `phd-runner`. This is a bit
unfortunate, as the command invocation accepts some arguments that are
consumed by the xtask command, and others that are passed through to
`phd-runner`. This commit adds [`allow_hyphen_values`] to the trailing
varargs in the Clap CLI for `cargo xtask phd`, so that the user doesn't
need to explicitly pass the `--` delimiter. This makes the CLI a bit
nicer to use.

[`allow_hyphen_values`]: https://docs.rs/clap/latest/clap/struct.Arg.html#method.allow_hyphen_values