PHD: Use ZFS clones for file-backed disks (#640)
When creating a file-backed disk from an artifact, create a ZFS snapshot and
of the dataset containing the artifact, then create and mount a clone of the
snapshot and back the disk from the clone. This is much, much less expensive
than copying the underlying file and markedly improves performance for test
runs with disks of nontrivial size. (On my local test machine, a full test run
with a 2 GiB Debian nocloud image took 688 seconds before this change and 610
seconds after, about a 10% speedup.)

This is unfortunately not likely to move the needle much on CI times because
the Alpine test image CI uses is very small. However, this should make it much
easier to switch to non-Alpine CI guest(s) without needlessly bloating CI times.

Tested with full runs against Debian 11 and Alpine guests.