Migrate LPC UART userspace pin states
Currently, interrupt pin state for userspace-emulated devices, like the
LPC UART, is accounted for both by the kernel in Bhyve and userspace in
Propolis. When userspace wishes to assert an IRQ for such a device, it
records the assertion of the IRQ line in userspace and then calls the
ioctl that asserts the IRQ in kernelspace. However, when such devices
are migrated, only the kernelspace Bhyve state is migrated, and any
userspace-tracked pin assertion counts are not. This means that these
states may become out of sync during a migration, potentially resulting
in a lost interrupt.

This branch introduces a mechanism for migrating userspace's
understanding of pin counts, in addition to migrating kernelspace's.
Because incrementing the userspace-tracked pin counts through the
IntrPin::assert or IntrPin::set_state methods may generate an
ioctl that changes the kernelspace state of that interrupt pin, we
cannot just call those methods while importing userspace pin assertion
counts. Instead, I've introduced a new import_state method to the
IntrPin trait, which is intended to be called only during imports.
This method sets the userspace-tracked pin state, the same as
set_state, but will not perform an ioctl call. This way, we can
import the userspace state independently of the kernelspace state,
without triggering additional interrupts that don't actually exist. I've
modified the LPC UART to use this new mechanism.

I've manually tested this change by running a migration between two lab
machines while mashing keys into the serial console, and the console
does not appear to get stuck.

Fixes #390