PHD: make callers consume serial buffer by hand
Change the semantics of waiting for characters in the serial buffer so that
waits return the preceding buffer contents but don't actually erase anything
from the buffer. Instead, make callers manually issue a "clear" command to
discard from the buffer anything they know they're not going to want, and update
the guest shell command sequences to clear the buffer between echoing the entire
command and pressing Enter to execute it.

This fixes issues with command echoing for Windows guests that don't redraw
characters that the guest believes are already on the screen--now those
characters don't get consumed from the buffer, so it's OK that the guest doesn't
explicitly reprint them.