PHD: add `pfexec` to xtask phd-runner invocation (#647)
Currently, when using `cargo xtask phd run` to run the PHD test suite,
ensuring that the runner has the correct permissions requires running
`pfexec cargo xtask phd`. This means that the commands run by the xtask
to set up the artifact store and temp directories, and to build
`phd-runner` and `propolis-server`, are run under the profile specified
by `pfexec`, resulting in a bunch of files in the Cargo `target/`
directory that the user doesn't actually own. This is annoying if you
want to then run `cargo clean` or something.

This commit changes `cargo xtask phd run` to just execute the runner
command with `pfexec` when compiled for Illumos. This saves the user
from having to type it themself, and (more importantly) means that the
binaries aren't built with a different owner than the user who builds
them.