smbios: add enums/bitflags types for some SMBIOS fields (#694)
Currently, most of the SMBIOS fields that we populate are magic numbers,
with comments describing what that magic number indicates. For SMBIOS
enum fields, this isn't too bad, but for bitfields like the various
"characteristics" fields, it's a bit of a shame, since the hex constant
doesn't always make it obvious which bit corresponds to which
characteristic.

To make this a bit nicer, this commit adds enums for most of the SMBIOS
enum fields we currently populate, and bitflags for most of the bitflags
fields. This way, we can refer to them by name rather than as magic
numbers. Also, I've made the `Default` implementation for the various
enum fields return the "unknown" (0x2) variant. Eventually, we may want
to default-initialize more SMBIOS tables --- but, we should probably
hold off on that until *all* the enum fields in the tables we populate
are Rust types with `Default` impls, so that we don't have to manually
initialize some of them to 0x2.

Furthermore, I've added `serde::{Serialize, Deserialize}`
implementations for these SMBIOS field types, in order to support
(eventually) overriding them from the config file. Using structured
types here also lets us perform some nicer validation of said
hypothetical future config file.

Current `dmidecode` output in an Alpine guest on this branch
matches what's on `master`.